requirements.txt
Flask
gTTS
pdfplumber
python-docx
python-dotenv
requests
google-generativeai